app/requirements.txt
transformers==4.46.3
ftfy==6.1.1
datasets==3.1.0
Pillow==9.5.0 
opencv-python==4.8.1.78
imgaug==0.4.0
gradio==4.38.1
accelerate==0.26.0
huggingface_hub==0.23.2
fastapi==0.112.4
tensorboard
qwen_vl_utils
image-reward
hpsv2
open-clip-torch
clip
segment_anything
openai
git+https://github.com/IDEA-Research/GroundingDINO.git